+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Each version section may have have subsections for: _Added_, _Changed_, _Removed
 ### Changed
 
 - `sockeye.train` now uses PyTorch's distributed data-parallel mode for multi-process (multi-GPU) training. Launch with: `torchrun --nproc_per_node N -m sockeye.train --dist ...`
+- Changed `--device-ids` argument (plural) to `--device-id` (singular). For multi-GPU training, see distributed mode noted above.
 - Updated default value: `--pad-vocab-to-multiple-of 8`
 - Removed `--horovod` argument used with `horovodrun` (use `--dist` with `torchrun`).
 - Removed `--optimizer-params` argument (use `--optimizer-betas`, `--optimizer-eps`).
@@ -31,6 +32,7 @@ Each version section may have have subsections for: _Added_, _Changed_, _Removed
 - Removed `--amp-scale-interval` argument.
 - Removed `--kvstore` argument.
 - Removed arguments: `--weight-init`, `--weight-init-scale` `--weight-init-xavier-factor-type`, `--weight-init-xavier-rand-type`
+- Removed `--decode-and-evaluate-device-id` argument.
 
 ## [2.3.22]
 ### Fixed